Changement sur un bug email verification

+++ routes/auth.py
@@ -61,17 +61,16 @@ def email_verification(tmp_string: str = None):
     if tmp_string:
         if doc := db.tmp_email_validation_url.find_one({"url": tmp_string}):
             # Drop tmp_user key of user session when link is GET by HTTP method
-            print(doc)
-            user = db.tmp_users.find_one({"email": doc["email"]})
-            print(user)
-            create_user(
-                user["username"],
-                user["promo"],
-                user["email"],
-                user["password"]
-            )
-            db.tmp_users.delete_one({"email": doc["email"]})
-            return render_template("email-verification.html", arg=tmp_string)
+            if user := db.tmp_users.find_one({"email": doc["email"]}):
+                create_user(
+                    user["username"],
+                    user["promo"],
+                    user["email"],
+                    user["password"]
+                )
+                db.tmp_users.delete_one({"email": doc["email"]})
+                return render_template("email-verification.html", arg=tmp_string)
+            return render_template("404.html")
         return render_template("404.html")
 
     return render_template("404.html")
